Fixed picky-compiler warning: make the cast from bool to int explicit.
svn path=/trunk/kdelibs/; revision=360181

+++ kstyles/keramik/gradients.cpp
@@ -46,7 +46,7 @@ namespace
 		
 		int key()
 		{
-			return m_menu ^ m_width ^ (m_height << 16) ^ ((m_color)<<8);
+			return (int)m_menu ^ m_width ^ (m_height << 16) ^ ((m_color)<<8);
 		}
 		
 		bool operator == (const GradientCacheEntry& other)
